Adding webpack 4
