+++ src/sass/style.css
@@ -2,8 +2,7 @@
 .custom-cookie-message-banner {
   width: 100%;
   position: fixed;
-  -webkit-transform: translateZ(0);
-          transform: translateZ(0);
+  transform: translateZ(0);
   will-change: transform;
   z-index: 2147483002;
   font-size: inherit;
@@ -26,12 +25,12 @@
 @media (min-width: 768px) {
   .custom-cookie-message-banner__content {
     flex-wrap: initial;
-    justify-content: centdder; }
+    justify-content: center; }
+
   .custom-cookie-message-banner__text {
     flex: initial;
     margin-bottom: 0;
     font-size: 100%; } }
-
 .custom-cookie-message-banner--bottom-fixed {
   bottom: 0; }
 
@@ -78,9 +77,9 @@
   .custom-cookie-message-banner__button {
     font-size: 100%;
     margin: 0 1.5em; }
+
   .custom-cookie-message-banner__content button {
     margin: 0 1.5em; } }
-
 .custom-cookie-message-banner__close {
   padding: 0 !important; }
 
@@ -108,8 +107,7 @@
   top: 0;
   left: 0;
   bottom: 0;
-  -webkit-transform: translateZ(0);
-          transform: translateZ(0);
+  transform: translateZ(0);
   will-change: transform;
   transition: all .3s ease 0s;
   width: 100%;
@@ -138,6 +136,7 @@
   max-height: 510px;
   max-width: 750px;
   background: white;
+  -webkit-box-shadow: 0 2px 6px 0 rgba(0, 0, 0, 0.15);
   box-shadow: 0 2px 6px 0 rgba(0, 0, 0, 0.15);
   border: 1px solid transparent; }
 
@@ -157,11 +156,9 @@
 @media screen and (min-width: 992px) {
   .custom-cookie-message-modal__title {
     font-size: 1.8em; } }
-
 @media screen and (min-width: 1200px) {
   .custom-cookie-message-modal__title {
     font-size: 2em; } }
-
 .custom-cookie-message-modal__close {
   position: absolute;
   top: -10px;
@@ -175,6 +172,7 @@
   line-height: 0;
   justify-content: center;
   align-items: center;
+  -webkit-box-shadow: 0 2px 6px 0 rgba(0, 0, 0, 0.15);
   box-shadow: 0 2px 6px 0 rgba(0, 0, 0, 0.15); }
 
 .custom-cookie-message-modal__close svg {
@@ -195,7 +193,6 @@
 @media (min-width: 515px) {
   .custom-cookie-message-modal__list {
     display: flex; } }
-
 .custom-cookie-message-modal__item {
   flex: 1;
   display: flex;
@@ -269,12 +266,14 @@
   .custom-cookie-message-banner,
   .custom-cookie-message-modal__header {
     padding: 0 1.5em; }
+
   .custom-cookie-message-modal__content {
     padding: 1.5em 1.5em 2em; }
+
   .custom-cookie-message-modal__item,
   .custom-cookie-message-modal__actions {
     padding: 1.5em; } }
-
 #intercom-container {
   z-index: 24999 !important; }
 
+/*# sourceMappingURL=style.css.map */
