+++ src/sass/style.scss
@@ -1,0 +1,1 @@
+@import "general";
